Move end_of_support.rs into acceptance.rs so tests are built faster
