+++ zebrad/tests/acceptance.rs
@@ -149,11 +149,14 @@ use color_eyre::{
 
 use zebra_chain::{
     block::{self, Height},
+    chain_tip::mock::MockChainTip,
     parameters::Network::{self, *},
 };
+use zebra_consensus::CheckpointList;
 use zebra_network::constants::PORT_IN_USE_ERROR;
 use zebra_node_services::rpc_client::RpcRequestClient;
 use zebra_state::constants::LOCK_FILE_ERROR;
+use zebrad::components::sync::end_of_support::{self, EOS_PANIC_AFTER, ESTIMATED_RELEASE_HEIGHT};
 
 use zebra_test::{args, command::ContextFrom, net::random_known_port, prelude::*};
 
@@ -2261,4 +2264,88 @@ async fn generate_checkpoints_mainnet() -> Result<()> {
 #[cfg(feature = "zebra-checkpoints")]
 async fn generate_checkpoints_testnet() -> Result<()> {
     common::checkpoints::run(Testnet).await
+}
+
+// Testing the end of support feature.
+
+// Estimated blocks per day with the current 75 seconds block spacing.
+const ESTIMATED_BLOCKS_PER_DAY: u32 = 1152;
+
+/// Test that the `end_of_support` function is working as expected.
+#[test]
+#[should_panic(expected = "Zebra refuses to run if the release date is older than")]
+fn end_of_support_panic() {
+    // We are in panic
+    let panic = ESTIMATED_RELEASE_HEIGHT + (EOS_PANIC_AFTER * ESTIMATED_BLOCKS_PER_DAY) + 1;
+
+    end_of_support::check(Height(panic), Network::Mainnet);
+}
+
+/// Test that the `end_of_support` function is working as expected.
+#[test]
+#[tracing_test::traced_test]
+fn end_of_support_function() {
+    // We are away from warn or panic
+    let no_warn = ESTIMATED_RELEASE_HEIGHT + (EOS_PANIC_AFTER * ESTIMATED_BLOCKS_PER_DAY)
+        - (30 * ESTIMATED_BLOCKS_PER_DAY);
+
+    end_of_support::check(Height(no_warn), Network::Mainnet);
+    assert!(logs_contain(
+        "Checking if Zebra release is inside support range ..."
+    ));
+    assert!(logs_contain("Zebra release is supported"));
+
+    // We are in warn range
+    let warn = ESTIMATED_RELEASE_HEIGHT + (EOS_PANIC_AFTER * 1152) - (3 * ESTIMATED_BLOCKS_PER_DAY);
+
+    end_of_support::check(Height(warn), Network::Mainnet);
+    assert!(logs_contain(
+        "Checking if Zebra release is inside support range ..."
+    ));
+    assert!(logs_contain(
+        "Your Zebra release is too old and it will stop running at block"
+    ));
+
+    // Panic is tested in `end_of_support_panic`
+}
+
+/// Test that we are never in end of support warning or panic.
+#[test]
+#[tracing_test::traced_test]
+fn end_of_support_date() {
+    // Get the list of checkpoints.
+    let list = CheckpointList::new(Network::Mainnet);
+
+    // Get the last one we have and use it as tip.
+    let higher_checkpoint = list.max_height();
+
+    end_of_support::check(higher_checkpoint, Network::Mainnet);
+    assert!(logs_contain(
+        "Checking if Zebra release is inside support range ..."
+    ));
+    assert!(!logs_contain(
+        "Your Zebra release is too old and it will stop running in"
+    ));
+}
+
+/// Check that the the end of support task is working.
+#[tokio::test]
+#[tracing_test::traced_test]
+async fn end_of_support_task() -> Result<()> {
+    let (latest_chain_tip, latest_chain_tip_sender) = MockChainTip::new();
+    latest_chain_tip_sender.send_best_tip_height(Height(10));
+
+    let eos_future = end_of_support::start(Network::Mainnet, latest_chain_tip);
+
+    let _ = tokio::time::timeout(Duration::from_secs(15), eos_future)
+        .await
+        .ok();
+
+    assert!(logs_contain(
+        "Checking if Zebra release is inside support range ..."
+    ));
+
+    assert!(logs_contain("Zebra release is supported"));
+
+    Ok(())
 }
